just finish the funtions charge and print

+++ HW2/IOThreadSave.cpp
@@ -65,7 +65,7 @@ IOThreadSave::IOThreadSave(){
 
 }
 
-void IOThreadSave::Bank_to_Log(int rate, int amount, int id) {
+void IOThreadSave::Bank_to_Log(float rate, int amount, int id) {
     pthread_mutex_lock(&(this->mutex_log));
     logFile<< "Bank: commisions of " << (rate * 100) << " % were charged, the bank gained " << amount << " from account " << id << endl;
     pthread_mutex_unlock(&(this->mutex_log));
